bread_chatbot/requirements.txt
gunicorn
uvicorn
blacknoise
boto3
gTTS
fastapi
openai
pydantic
requests
python-dotenv